+++ src/test/kotlin/me/bossm0n5t3r/leetcode/uniquePaths2/UniquePathsTest2.kt
@@ -1,10 +1,10 @@
-package uniquePaths2
+package me.bossm0n5t3r.leetcode.uniquePaths2
 
 import org.junit.jupiter.api.Test
 import kotlin.test.assertEquals
 
 class UniquePathsTest2 {
-    private val uniquePaths2 = UniquePaths2()
+    private val uniquePaths2 = UniquePaths2.Solution()
 
     data class UniquePaths2TestData(
         val obstacleGrid: Array<IntArray>,
